requirements.txt
torch
numpy
opencv-python
mss
cmake
pynput
pyautogui
stable-baselines3
stable-retro
